[trivia] Re-add diep.io
However this time it will be restricted to a few guilds. As Chiaki
gets bigger, less people will know or care what diep.io is, and we
shouldn't burden the user with that. However, we can't alienate Chiaki's
roots either. She started with diep.io, so we still need the category
there.