Fix random KeyErrors when unloading extensions.
The bug is a bit complicated. It turns out ChainMap.__iter__ copies all
the keys before iterating. The issue is that MutableMapping.items (the
base class for ChainMap) calls __iter__ and then does (key, mapping[key]).

The only problem is we were deleting keys in the alias mapping. Causing
random errors. In order to fix this, Chiaki.cogs is now just a CIDict
containing all the cogs rather than a ChainMap.