Use two columns to list commands in help
This gives the major advantage of there being way less pages. We started
from 30 to 16 thanks to grouping them up in big categories. (de1da4f).

However, this was still too many pages. Now thanks to this one, we now
only need 8 pages, while still having the pages be reasonably compact.

Of course ideally we'd cut pagination altogether but this is the next
best thing due to Chiaki having way too many commands at the moment.

The only downside is that aliases are ignored but that's not too much of
a loss since the decrease in pages to go through is worth it.