Give fallbacks to commands that don't necessarily need embeds
This was mainly a judgement call, but I decided that for a few commands
where embeds weren't the main focus, that they shouldn't have to require
Embed Links.

Of course this is only a few commands, because the rest are either too
embedded to figure out a fallback, or never used embeds in the first
place.

Make a fallback for ->invite