Remove destination arg in BaseReactionPaginator.interact
Originally I added this because I thought I needed to specify an
alternate destination at some point. Many months later, I still haven't
done that.

To make matters worse, it made it hard to check the destination to see
if Chiaki had the perms to embeds or add reactions. So I just removed it.

If I really need it, I'll find a way to include it