Replace the first page with the Category List
Today's a sad day. The beloved first page with Chiaki playing with a
Game Girl has been removed.

This decision didn't come easily. However after weeks of debating, it
has been deemed too purely aesthetic and style-over-substance to be kept.

There was a plan to give it an easter egg but it's not really worth it
as it would complicate the code too much.

Also added something about ->help command