Hide the concept of "cogs" from the end user
The average user won't know or care about what "cogs" are, cogs are only
good for separating commands into separate files. Because of this, there
are way too many cogs (around 30 as of this moment), and some of them
have just one or two commands, far too much for the user to keep track.