Force a timeout of 2 minutes in ->10s
The real reason why Chiaki hangs so much is because a bunch of people
tended to leave active 10s games running. Because they can go on
indefinitely, and I forgot to release the DB connection on them, they
effectively permanently acquire a connection, causing hangs to happen.

While technically can be considered breaking, they should not be lasting
this long anyways. After all, it's a test of 10 seconds, not a test of
how long it can go.