Clean up how statuses are formatted
When I first made this I ran under the assumption that the config file
wouldn't have the emojis. However, I didn't account for if the bot
didn't have the Use External Emojis permission.

It's easier to assume the config file will have valid emojis, than to
assume that the bot would have the External Emojis permission.