Add ->prefix set
Previously, in order to change the prefix, one would have to do:
1. ->prefix add <new-prefix>
2. ->prefix remove <old-prefix>

This was rather verbose for a simple prefix command, and more often than
not, people just want to *set* a prefix, rather than add one.

This makes it easier to just simply set a custom prefix, as it's a
common thing on various other bots.

Note that ->prefix <new-prefix> should not be done because of the
subcommands. A typo when using the subcommands would destroy the whole
server prefixes config, because prefix setting replaces the whole thing
with the new prefix.