Replace SessionManager with normal dict
SessionManager was a redundant abstraction as it only was needed
to create temporary sessions.

Cogs also now have the same attribute name (sessions) for the
currently running games.

Finally IDs are used where possible instead of storing the
whole object as the key to avoid too many references to
the channel object.